ft_printf error correction

+++ ft_print_hex.c
@@ -6,13 +6,13 @@
 /*   By: linuxusr <linuxusr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/11 18:43:51 by linuxusr          #+#    #+#             */
-/*   Updated: 2021/10/11 18:50:15 by linuxusr         ###   ########.fr       */
+/*   Updated: 2021/10/13 14:15:22 by linuxusr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-size_t	ft_print_hex(ssize_t n)
+size_t	ft_print_hex(unsigned int n)
 {
 	char	*s;
 	size_t	len;
