+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: linuxusr <linuxusr@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/04 17:34:18 by linuxusr          #+#    #+#             */
-/*   Updated: 2021/10/11 18:49:20 by linuxusr         ###   ########.fr       */
+/*   Updated: 2021/10/13 14:28:23 by linuxusr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,8 +21,8 @@ size_t	ft_print_str(char *s);
 size_t	ft_print_ptr(void *p);
 size_t	ft_print_int(int n);
 size_t	ft_print_uint(unsigned int u);
-size_t	ft_print_hex(ssize_t n);
-size_t	ft_print_uhex(ssize_t n);
+size_t	ft_print_hex(unsigned int n);
+size_t	ft_print_uhex(unsigned int n);
 int		ft_printf(const char *s, ...);
 
 #endif
